Redesign 'show-storage-struct' to 'db struct'
