+++ CMakeLists.txt
@@ -46,8 +46,6 @@ list (APPEND WSJCPP_SOURCES "src/core/export_libwsjcppcli_java_android.h")
 list (APPEND WSJCPP_SOURCES "src/core/export_libwsjcppcli_java_android.cpp")
 list (APPEND WSJCPP_SOURCES "src/core/export_libwsjcppcli_py.h")
 list (APPEND WSJCPP_SOURCES "src/core/export_libwsjcppcli_py.cpp")
-list (APPEND WSJCPP_SOURCES "src/core/export_struct_of_storage.h")
-list (APPEND WSJCPP_SOURCES "src/core/export_struct_of_storage.cpp")
 list (APPEND WSJCPP_SOURCES "src/core/fallen.h")
 list (APPEND WSJCPP_SOURCES "src/core/fallen.cpp")
 list (APPEND WSJCPP_SOURCES "src/core/jobs_pool.h")
