+++ src/argument_processors/argument_processor_database.h
@@ -6,14 +6,19 @@
 class ArgumentProcessorDatabase : public WsjcppArgumentProcessor {
     public:
         ArgumentProcessorDatabase();
-
         virtual int exec(const std::vector<std::string> &vRoutes, const std::vector<std::string> &vSubParams) override;
 };
 
 class ArgumentProcessorDatabaseTestConnection : public WsjcppArgumentProcessor {
     public:
         ArgumentProcessorDatabaseTestConnection();
+        virtual int exec(const std::vector<std::string> &vRoutes, const std::vector<std::string> &vSubParams) override;
+};
 
+
+class ArgumentProcessorDatabaseStruct : public WsjcppArgumentProcessor {
+    public:
+        ArgumentProcessorDatabaseStruct();
         virtual int exec(const std::vector<std::string> &vRoutes, const std::vector<std::string> &vSubParams) override;
 };
