+++ src/argument_processors/argument_processor_main.h
@@ -21,7 +21,6 @@ class ArgumentProcessorMain : public WsjcppArgumentProcessor {
 class ArgumentProcessorVersion : public WsjcppArgumentProcessor {
     public:
         ArgumentProcessorVersion();
-
         virtual int exec(const std::vector<std::string> &vRoutes, const std::vector<std::string> &vSubParams) override;
 };
