--- src/core/export_struct_of_storage.h
@@ -1,10 +1,0 @@
-#ifndef EXPORT_STRUCT_OF_STORAGE_H
-#define EXPORT_STRUCT_OF_STORAGE_H
-
-class ExportStructOfStorage {
-    public:
-        static void print();
-
-};
-
-#endif // EXPORT_STRUCT_OF_STORAGE_H
